+++ src/main/java/ar/edu/unahur/obj2/energia/SistemasPerifericos/SistemaInteresado.java
@@ -4,5 +4,5 @@ import ar.edu.unahur.obj2.energia.Bateria.Bateria;
 
 public interface SistemaInteresado {
 
-    void notificar(Bateria bateria, String tipoMovimiento, double cantidad);
+    void notificar(Bateria bateria, String tipoMovimiento, Integer cantidad);
 }
